Hotel Management System
In the world of hotels and resorts, the competition in the market is very high. Every year, due to poor management, many hotels and resorts become stale and go out of competition. Managements the key factor in determining whether the hotel will stay in the competition or run out of business. First impression of a hotel to the public is its management, which, if it is upto the mark, raises the reputation of the hotels in the eyes of the people. In today’s world, the only way to keep management up to the mark is to make labor and technology work in compliment with each other. Most of the hotels today use a labor-based management system, which, as a result causes the management to lack behind the demands of people as there are somany people checking in and out of each hotel every single day that the labor based management system cannot keep up and then the complains of people rises which degrades the reputation of the hotel and eventually it goes out of business. On the other hand, very heavy technology-based management systems, which many 5 star and 7-star hotels use, fail to cater to the needs of the people and provide extra care to them, which also downgrades the relevancy of the hotel.